Move tests to separate project
