+++ paparazzi-gradle-plugin/src/test/projects/verify-animation-gif/src/main/res/drawable/animated_vector_rotation.xml
