+++ paparazzi-gradle-plugin/src/test/projects/verify-animation-gif/src/main/res/drawable/vector_rotation_base.xml
