+++ paparazzi-gradle-plugin/src/test/projects/verify-animation-gif/src/test/snapshots/videos/app.cash.paparazzi.plugin.test_AnimatedVectorDrawableTest_rotatesAcrossFrames_rotation.png
